requirements.txt
praw 

pyspark==3.4.1

pymongo

confluent-kafka

python-dotenv